% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_holidays_short_term.R
\name{add_holidays_short_term}
\alias{add_holidays_short_term}
\title{Title}
\usage{
add_holidays_short_term(shortterm)
}
\arguments{
\item{shortterm}{}
}
\description{
Title
}
\examples{
shortterm_holidays_example <- add_holidays_short_term(decomposed_load_example$shortterm)
shortterm_holidays_example[,c(1,2,11)]
}
