% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_load_data.R
\name{decompose_load_data}
\alias{decompose_load_data}
\title{Decomposition into long-, mid- and short-term component}
\usage{
decompose_load_data(load_data)
}
\arguments{
\item{load_data}{A dataframe object with "load", "date", "unit", and "country" columns}
}
\value{
A list of three dataframes with long-term trend, mid-term seasonality, short-term seasonality
}
\description{
This function decomposes the load data into three components: a yearly long-term trend, a daily mid-term seasonality, and an hourly short-term seasonality. If the data is available only at a daily resolution, the calculation of hourly seasonality is skipped. The results of the decomposition are returned as a list of dataframes.
}
\examples{
decomposed_load_example <- decompose_load_data(no_missing_data_example)
}
