% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_demand_data}
\alias{example_demand_data}
\title{Example Demand Data}
\format{
A data frame with 43,769 rows and 7 columns:
\describe{
\item{\code{date}}{The date and time of the demand measurement (in \verb{YYYY-MM-DD HH:MM:SS} format).}
\item{\code{load}}{The electricity demand.}
\item{\code{unit}}{The unit of measurement (MW).}
\item{\code{year}}{The year of the respective timepoint.}
\item{\code{time_interval}}{The time interval at which the demand was reported (e.g., \verb{60 mins}).}
\item{\code{country}}{The country, represented by the ISO2C country code (e.g., \code{FR} for France).}
\item{\code{example}}{A boolean indicator to mark this dataset as an example dataset.}
}
}
\source{
Transparency Platform of the European Network of
Transmission System Operators for Electricity (ENTSO-E, https://transparency.entsoe.eu/)
}
\usage{
example_demand_data
}
\description{
This dataset contains the hourly electricity demand data of France from 2017 until 2021.
}
\keyword{dataset}
\keyword{demand}
\keyword{electricity}
