% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_load_data}
\alias{example_load_data}
\title{Example Dataset}
\format{
A data frame with \code{n} rows and \code{m} columns:
\describe{
\item{\code{column1}}{Description of column1.}
\item{\code{column2}}{Description of column2.}
\item{\code{...}}{And so on for all your columns.}
}
}
\source{
Describe the source of your data, if applicable.
}
\usage{
example_load_data
}
\description{
A brief description of your dataset. Explain what it represents, how it was collected,
and any important variables it contains.
}
\keyword{datasets}
