% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_longterm_data}
\alias{example_longterm_data}
\title{Example Longterm Data}
\format{
A data frame with 16 rows and 4 columns:
\describe{
\item{\code{country}}{The country, represented by the ISO2C country code (e.g., \code{FR} for France).}
\item{\code{year}}{The year of the observation.}
\item{\code{avg_hourly_demand}}{The average hourly electricity demand (in megawatts) for each year.}
\item{\code{example}}{A boolean indicator to mark this dataset as an example dataset.}
}
}
\source{
Transparency Platform of the European Network of
Transmission System Operators for Electricity (ENTSO-E, https://transparency.entsoe.eu/) ; ENTSO-E Power Stats Archive (https://www.entsoe.eu/data/power-stats/)
}
\usage{
example_longterm_data
}
\description{
This dataset contains the average hourly electricity demand per year for France from 2006 until 2021.
It is an extension of the long-term component of \code{\link{example_decomposed_data}} with historical data from the ENTSO-E Power Stats archive (https://www.entsoe.eu/data/power-stats/).
}
\keyword{dataset}
\keyword{longterm}
