% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_data.R
\name{fill_missing_data}
\alias{fill_missing_data}
\title{Replace missing values in the load data set}
\usage{
fill_missing_data(load_data)
}
\arguments{
\item{load_data}{Data Frame with load data. Data Frame must contain the following columns:
"date" consisting of the datetime values,
"load" consisting of the load values,
"unit" indicating the measured unit (e.g. MW),
"country" indicating the country's ISO2C code}
}
\value{
Data Frame with filled up load values, date, unit, year, time resolution, ISO2C Country Code
}
\description{
The purpose of this function is to substitute missing values with the corresponding values from the dataset, exactly one week prior at the same time point. For example, if there is no load value available for May 12th at 20:00, the value recorded on May 7th at 20:00 will be used as a replacement.
This function is primarily designed to handle minor gaps in the dataset acquired using the "get_entsoE_data()" function. To use this function with other datasets, it is important that the input data frame adheres to the required column naming conventions.
}
\examples{
nrow(example_load_data)
example_load_data[858:862,]
no_missing_data_example <- fill_missing_data(example_load_data)
nrow(no_missing_data_example)
no_missing_data_example[858:876,]
}
