% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entsoE_data.R
\name{get_entsoE_data}
\alias{get_entsoE_data}
\title{Title}
\usage{
get_entsoE_data(
  start_year,
  end_year,
  country,
  api_key = "5ca5937c-7eae-4302-b444-5042ab55d8ef"
)
}
\arguments{
\item{api_key}{}
}
\description{
Title
}
