% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historic_load_data.R
\name{get_historic_load_data}
\alias{get_historic_load_data}
\title{Adding historic yearly load average data}
\usage{
get_historic_load_data(longterm)
}
\arguments{
\item{longterm}{A dataframe object with "country", "year", and "avg_hourly_demand" columns}
}
\value{
Data Frame with "country", "year", and "avg_hourly_demand"
}
\description{
This function utilizes two historical load datasets obtained from the archive of https://www.entsoe.eu/data/power-stats/.
It calculates the hourly average load for each year and appends the results to the long-term dataframe generated by the decompose_load_data() function.
}
\examples{
print("Input dataframe should be in the following form:")
decomposed_load_data_example$longterm
long_term_example <- get_historic_load_data(decomposed_load_data_example$longterm)
long_term_example
}
