% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_data.R
\name{get_weather_data}
\alias{get_weather_data}
\title{Title}
\usage{
get_weather_data(midterm)
}
\arguments{
\item{midterm}{}
}
\description{
Title
}
\examples{
midterm_all_example <- get_weather_data(midterm_holidays_example)
midterm_all_example$midterm
midterm_all_example$temperature_data
}
