% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historic_entsodata1}
\alias{historic_entsodata1}
\title{Historical load data from the ENTSO-E Powerstats Archive}
\format{
A data frame with 116,805 rows and 29 columns:
\describe{
\item{\code{Hourly load values 2006-2015}}{This column contains the respective country name, represented by the ISO2C country code (e.g., \code{FR} for France). }
\item{\code{...2}}{The respective year.}
\item{\code{...3}}{The respective month.}
\item{\code{...4}}{The respective day.}
\item{\code{...5}}{The coverage ratio. To get the real value, you need to scale them to 100\% for coverage ratio less than 100.}
\item{\code{...6},\code{...7},\code{...8},\code{...9},\code{...10}, \code{...11}, \code{...12}, \code{...13}, \code{...14}, \code{...15}, \code{...16}, \code{...17}, \code{...18}, \code{...19}, \code{...20}, \code{...21}, \code{...22}, \code{...23}, \code{...24}, \code{...25}, \code{...26}, \code{...27}, \code{...28}, \code{...29}}{The hourly load value at the respective time of day, starting from 0:00-0:59 until 23:00-23:59.}
}
}
\source{
ENTSO-E Power Stats Archive (https://www.entsoe.eu/data/power-stats/)
}
\usage{
historic_entsodata1
}
\description{
This dataset contains historic hourly load data from 2006 until 2015 for various European countries.
}
\keyword{data}
\keyword{historic}
\keyword{power}
