% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historic_entsodata2}
\alias{historic_entsodata2}
\title{Historical load data from the ENTSO-E Powerstats Archive}
\format{
A data frame with 631600 rows and 9 columns:
\describe{
\item{\code{MeasureItem}}{Describes what is measured (Monthly Hourly Load Values)}
\item{\code{DateUTC}}{The datetime in UTC format (\verb{YYYY-MM-DD hh:mm:ss})}
\item{\code{DateShort}}{The date in UTC format (\code{YYYY-MM-DD})}
\item{\code{TimeFrom}}{The starting time of the observation period.}
\item{\code{TimeTo}}{The ending time of the observation period.}
\item{\code{CountryCode}}{The ISO2C country code (e.g., \code{FR} for France).}
\item{\code{Cov_ratio}}{The coverage ratio. For coverage ratios less than 100\%, scale to 100\% to obtain the real value.}
\item{\code{Value}}{The measured value. Scale it to 100\% for coverage ratios less than 100\%.}
\item{\code{Value_ScaleTo100}}{The scaled value adjusted to a 100\% coverage ratio.}
}
}
\source{
ENTSO-E Power Stats Archive (https://www.entsoe.eu/data/power-stats/)
}
\usage{
historic_entsodata2
}
\description{
This dataset contains historic hourly load data from 2015 until 2017 for various European countries.
}
\keyword{data}
\keyword{historic}
\keyword{power}
