% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_future.R
\name{long_term_future}
\alias{long_term_future}
\title{Title}
\usage{
long_term_future(longterm_predictions, end_year)
}
\arguments{
\item{longterm_predictions}{Dataframe. Generated by \code{\link{long_term_future_data}}}

\item{end_year}{Integer. Specifies the final year for which future predictions will be generated.}
}
\value{
The extended initial dataframe with the future predictions for each of the 3 best long term models.
}
\description{
Title
}
\examples{
\dontrun{
longterm_future_predictions <- long_term_future(longterm_future_data_example,end_year=2028)
}
}
\seealso{
See also function \code{\link{mid_term_future}} and \code{\link{short_term_future}} for the other prediction models and \code{\link{long_term_future_data}} for the covariate download.
}
