% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_future_data.R
\name{long_term_future_data}
\alias{long_term_future_data}
\title{Title}
\usage{
long_term_future_data(longterm_predictions, end_year, dataset = "WEO")
}
\arguments{
\item{longterm_predictions}{Dataframe. Generated by \code{\link{long_term_lm}}}

\item{end_year}{Integer. Specifies the final year for which future predictions will be generated.}

\item{dataset}{Character. By default the World Economic Outlook (WEO) Database April 2023 edition is used to generate covariate predictions for the long term models. If the dataset option is set to anything else than "WEO" the function will tell you for which covariates data will be needed to run \code{\link{long_term_future}}.}
}
\value{
The extended initial dataframe until the specified end year with the covariate predictions if dataset="WEO" or with empty values and a note for which covariates forecasts must be supplied.
}
\description{
Title
}
