% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_holidays_short_term.R
\name{orakle.add_holidays_short_term}
\alias{orakle.add_holidays_short_term}
\title{Title}
\usage{
orakle.add_holidays_short_term(shortterm)
}
\arguments{
\item{shortterm}{}
}
\description{
Title
}
