% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_lm.R
\name{orakle.long_term_lm}
\alias{orakle.long_term_lm}
\title{Title}
\usage{
orakle.long_term_lm(
  longterm_all_data,
  training_set_ratio = 0.1,
  testquant = 500
)
}
\arguments{
\item{testquant}{}
}
\description{
Title
}
