% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_lm.R
\name{orakle.shortterm_lm}
\alias{orakle.shortterm_lm}
\title{Title}
\usage{
orakle.shortterm_lm(short_term_data, training_set_ratio = 0.2)
}
\arguments{
\item{training_set_ratio}{}
}
\description{
Title
}
