% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weo_data}
\alias{weo_data}
\title{World Economic Outlook (WEO) Data}
\format{
A data frame with 980 rows and 59 columns:
\describe{
\item{\code{WEO.Country.Code}}{Unique code assigned to each country by the WEO database.}
\item{\code{ISO}}{The ISO3 country code (e.g., \code{FRA} for France).}
\item{\code{WEO.Subject.Code}}{Unique code representing the economic indicator or subject in the WEO database.}
\item{\code{Country}}{The name of the country.}
\item{\code{Subject.Descriptor}}{Description of the economic indicator.}
\item{\code{Subject.Notes}}{Additional notes or details about the economic indicator.}
\item{\code{Units}}{The unit of measurement for the indicator (e.g., percentage, persons, national currency).}
\item{\code{Scale}}{The scaling factor for the indicator values (e.g., "Billions").}
\item{\code{Country.Series.specific.Notes}}{Country-specific notes about the data series.}
\item{\code{X1980}, \code{X1981}, \code{X1982}, \code{X1983}, \code{X1984},
\code{X1985}, \code{X1986}, \code{X1987}, \code{X1988}, \code{X1989},
\code{X1990}, \code{X1991}, \code{X1992}, \code{X1993}, \code{X1994},
\code{X1995}, \code{X1996}, \code{X1997}, \code{X1998}, \code{X1999},
\code{X2000}, \code{X2001}, \code{X2002}, \code{X2003}, \code{X2004},
\code{X2005}, \code{X2006}, \code{X2007}, \code{X2008}, \code{X2009},
\code{X2010}, \code{X2011}, \code{X2012}, \code{X2013}, \code{X2014},
\code{X2015}, \code{X2016}, \code{X2017}, \code{X2018}, \code{X2019},
\code{X2020}, \code{X2021}, \code{X2022}, \code{X2023}, \code{X2024},
\code{X2025}, \code{X2026}, \code{X2027}, \code{X2028}}{The annual value of the economic indicator for the respective year, starting from 1980 (\code{X1980}) to 2028 (\code{X2028}).}
\item{\code{Estimates.Start.After}}{The year after which data values are based on projections instead of past values.}
}
}
\source{
World Economic Outlook Database (April 2023 edition) of the International Monetary Fund (IMF) (https://www.imf.org/en/Publications/WEO/weo-database/2023/October)
}
\usage{
weo_data
}
\description{
This dataset contains macroeconomic data and projections from the International Monetary Fund (IMF) World Economic Outlook (WEO). It includes annual data for multiple countries and economic indicators, spanning from 1980 to 2028.
}
\keyword{IMF}
\keyword{WEO}
\keyword{dataset}
\keyword{macroeconomics}
\keyword{projections}
