% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_holidays_mid_term.R
\name{add_holidays_mid_term}
\alias{add_holidays_mid_term}
\title{Title}
\usage{
add_holidays_mid_term(midterm)
}
\arguments{
\item{midterm}{}
}
\description{
Title
}
\examples{
midterm_holidays_example <- add_holidays_mid_term(decomposed_load_example$midterm)
midterm_holidays_example
}
