% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_holidays_mid_term.R
\name{add_holidays_mid_term}
\alias{add_holidays_mid_term}
\title{Add holidays to the mid-term series}
\usage{
add_holidays_mid_term(midterm_data)
}
\arguments{
\item{midterm_data}{The mid-term data series resulting from the function \code{\link{decompose_load_data}}.}
}
\value{
The mid-term series with an additional column of holiday dummies.
}
\description{
This function adds a dummy variable for holidays to the mid-term data series. Information on the holidays is retrieved from "https://date.nager.at/api/v3/publicholidays/".
}
\examples{
\dontrun{
example_midterm_demand_data <- add_holidays_mid_term(example_decomposed_data$midterm)
example_midterm_demand_data
}
}
\seealso{
See also \code{\link{mid_term_lm}} for the prediction model.
}
