% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_holidays_short_term.R
\name{add_holidays_short_term}
\alias{add_holidays_short_term}
\title{Add holidays to the short-term series}
\usage{
add_holidays_short_term(shortterm)
}
\arguments{
\item{shortterm}{The short-term data series resulting from \code{\link{decompose_load_data}}}
}
\value{
The short-term series with an additional column of holiday dummies.
}
\description{
This function adds a dummy variable for holidays to the short-term data series. Information on the holidays is retrieved from "https://date.nager.at/api/v3/publicholidays/".
}
\examples{
example_shortterm_demand_data <- add_holidays_short_term(example_decomposed_data$shortterm)
example_shortterm_demand_data[1:5, c(1, 2, 11)]
}
