% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_models.R
\name{combine_models}
\alias{combine_models}
\title{Title}
\usage{
combine_models(
  longterm_all_data_predicted,
  midterm_all_data_predicted,
  short_term_data_predicted,
  longterm_model_number = 1
)
}
\arguments{
\item{short_term_data_predicted}{}
}
\description{
Title
}
