% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_models.R
\name{combine_models}
\alias{combine_models}
\title{Combine forecast models}
\usage{
combine_models(
  longterm_all_data_predicted,
  midterm_all_data_predicted,
  short_term_data_predicted,
  longterm_model_number = 1
)
}
\arguments{
\item{longterm_all_data_predicted}{Dataframe. The object resulting from function \code{\link{long_term_lm}}.}

\item{midterm_all_data_predicted}{Dataframe. The object resulting from function \code{\link{mid_term_lm}}.}

\item{short_term_data_predicted}{Dataframe. The object resulting from function \code{\link{short_term_lm}}.}

\item{longterm_model_number}{Integer. Specifies which of the 3 best long-term models should be used.}
}
\value{
The combined model results.
}
\description{
This function combines the three separate forecasts for the low, mid and high frequency model. The three separate forecasts need to be run first.
}
\examples{
\dontrun{
combined_model_results <- long_term_lm(longterm_all_data_example,
test_set_steps=2,testquant = 500)
}

}
