% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_models.R
\name{combine_models}
\alias{combine_models}
\title{Combine forecast models}
\usage{
combine_models(
  longterm_predictions,
  midterm_predictions,
  shortterm_predictions,
  longterm_model_number = 1
)
}
\arguments{
\item{longterm_predictions}{Dataframe. The object resulting from function \code{\link{long_term_lm}}.}

\item{midterm_predictions}{Dataframe. The object resulting from function \code{\link{mid_term_lm}}.}

\item{shortterm_predictions}{Dataframe. The object resulting from function \code{\link{short_term_lm}}.}

\item{longterm_model_number}{Integer. Specifies which of the 3 best long-term models should be used.}
}
\value{
Dataframe. The combined model results.
}
\description{
This function combines the three separate forecasts for the low, mid and high frequency model. The three separate forecasts need to be run first.
}
\examples{
example_full_model_predictions <- combine_models(example_longterm_predictions,
example_midterm_predictions,example_shortterm_predictions,longterm_model_number =1)

}
