% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_models_future.R
\name{combine_models_future}
\alias{combine_models_future}
\title{Combine forecast models for future predictions}
\usage{
combine_models_future(
  longterm_future_predictions,
  midterm_future_predictions,
  shortterm_future_predictions,
  longterm_model_number = 1
)
}
\arguments{
\item{longterm_future_predictions}{Dataframe. The object resulting from function \code{\link{long_term_future}}.}

\item{midterm_future_predictions}{Dataframe. The object resulting from function \code{\link{mid_term_future}}.}

\item{shortterm_future_predictions}{Dataframe. The object resulting from function \code{\link{short_term_future}}.}

\item{longterm_model_number}{Integer. Specifies which of the 3 best long-term models should be used.}
}
\value{
Dataframe. The combined model results.
}
\description{
This function combines the three separate future forecasts for the low, mid and high frequency model. The three separate forecasts need to be run first and should have the same end_year.
}
\examples{
example_full_model_future_predictions <- combine_models_future(example_longterm_future_predictions,
  example_midterm_future_predictions, example_shortterm_future_predictions,
  longterm_model_number = 1
)
}
\seealso{
See also functions \code{\link{long_term_future}}, \code{\link{mid_term_future}}, and \code{\link{short_term_future}} for the prediction models.
}
