% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_load_data.R
\name{decompose_load_data}
\alias{decompose_load_data}
\title{Decomposing the load data into long-, mid- and short-term component}
\usage{
decompose_load_data(load_data)
}
\arguments{
\item{load_data}{A dataframe object with "load", "date", "unit", and "country" columns
\item{load} consisting of the load values,
\item{date} consisting of the datetime values,
\item{unit} indicating the measured unit (e.g. MW),
\item{country} indicating the country's ISO2C code}
}
\value{
A list of three dataframes with
\item{longterm} dataframe of long-term trend with country, year and yearly average hourly demand.
\item{midterm} datafram of mid-term component with country, date, year, month, day, week day, average hourly demand and seasonal average hourly demand.
\item{shortterm} dataframe of short-term component with country, date, year, month, day, week day, hour, hourly demand and hourly demand trend and trend and season corrected.
}
\description{
This function decomposes the load data into three components: a yearly long-term trend, a daily mid-term seasonality, and an hourly short-term seasonality. If the data is available only at a daily resolution, the calculation of hourly seasonality is skipped. The results of the decomposition are returned as a list of dataframes. The series are plotted additionally.
}
\examples{
decomposed_load_example <- decompose_load_data(no_missing_data_example)
}
