% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_load_data.R
\name{decompose_load_data}
\alias{decompose_load_data}
\title{Trend and seasonality decomposition}
\usage{
decompose_load_data(load_data)
}
\arguments{
\item{load_data}{A dataframe object with load,date,unit and country columns}
}
\value{
list of three dataframes with long-term trend, mid-term seasonality, short-term seasonality
}
\description{
Decomposes the load data into a yearly long-term, trend a daily mid-term seasonality and an hourly short-term
seasonality. If the data is only in daily resolution no hourly seasonality is calculated. The results are returned as a list of dataframes.
}
\examples{
a <- 5
}
