% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_data.R
\name{fill_missing_data}
\alias{fill_missing_data}
\title{Replace missing values in the load data set}
\usage{
fill_missing_data(load_data)
}
\arguments{
\item{load_data}{Data Frame with load data. Data Frame must contain the following columns:
\item{date} consisting of the datetime values,
\item{load} consisting of the load values,
\item{unit} indicating the measured unit (e.g. MW),
\item{country} indicating the country's ISO2C code}
}
\value{
Data Frame with completed load values, date, unit, year, time resolution, ISO2C Country Code
}
\description{
This function substitutes missing values with the corresponding values at the same hour exactly one week prior. For example, if there is no load value available for May 12th at 20:00, the value recorded on May 7th at 20:00 will be used as a replacement.
This function is primarily designed to handle minor gaps in the dataset acquired using the "get_entsoE_data()" function. To use this function with other datasets, it is important that the input data frame adheres to the required column naming conventions.
}
\examples{
library(ggplot2)
no_missing_data_example <- fill_missing_data(example_load_data)
example_df<- as.data.frame(seq.POSIXt(example_load_data$date[841],example_load_data$date[870],"hour"))
example_df$before <- NA
example_df$before[example_df[,1] \%in\% example_load_data$date] <- example_load_data$load[example_load_data$date \%in\% example_df[,1]]
example_df$after <-   no_missing_data_example$load[no_missing_data_example$date \%in\% example_df[,1]]
ggplot(example_df,aes(x=example_df[,1]))+ geom_line(aes(y=after,colour="after data filling"))+
geom_line(aes(y=before,colour="before data filling"))+
 xlab("\nHour")+ylab("Load [MW]\n")+theme(legend.title = element_blank())+
 scale_x_continuous(breaks = c(example_df[1,1],example_df[25,1]),labels = c(as.Date(example_df[1,1]),as.Date(example_df[25,1])))
}
