% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entsoE_data.R
\name{get_entsoE_data}
\alias{get_entsoE_data}
\title{Load Data Acquisition from the ENTSO-E Transparency Platform}
\usage{
get_entsoE_data(
  start_year,
  end_year,
  country,
  api_key = "5ca5937c-7eae-4302-b444-5042ab55d8ef"
)
}
\arguments{
\item{start_year}{The starting year for which load data will be requested.}

\item{end_year}{The final year for which load data will be requested.}

\item{country}{The country name for which load data will be requested.}

\item{api_key}{A valid API key for the ENTSO-E Transparency Platform. If none is provided one of the deposited keys will be used.}
}
\value{
Data Frame with the Date, Load, Unit, Year, Time Resolution, ISO2C Country Code
}
\description{
Makes various API Requests to the Transparency Platform of the European Network of
Transmission System Operators for Electricity (ENTSO-E) and stores the results in a data frame. https://transparency.entsoe.eu/
}
\examples{
get_entsoE_data(2017,2021,"France")

}
