% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_future.R
\name{long_term_future}
\alias{long_term_future}
\title{Long-term trend predictions for future years}
\usage{
long_term_future(longterm_future_macro_data)
}
\arguments{
\item{longterm_future_macro_data}{Dataframe. Generated by \code{\link{long_term_future_data}}}
}
\value{
The extended initial dataframe with the future predictions for each of the 3 best long term models.
}
\description{
Long-term trend predictions for future years
}
\examples{
\dontrun{
example_longterm_future_predictions <- long_term_future(example_longterm_future_macro_data)
}
}
\seealso{
See also function \code{\link{mid_term_future}} and \code{\link{short_term_future}} for the other prediction models and \code{\link{long_term_future_data}} for the covariate download.
}
