% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_future.R
\name{long_term_future}
\alias{long_term_future}
\title{Title}
\usage{
long_term_future(longterm_predictions, end_year)
}
\arguments{
\item{longterm_predictions}{Dataframe. Generated by long_term_lm()}

\item{end_year}{Integer. Specifies the final year for which future predictions will be generated.}
}
\description{
Title
}
