% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_future_data.R
\name{long_term_future_data}
\alias{long_term_future_data}
\title{Title}
\usage{
long_term_future_data(longterm_predictions, end_year)
}
\arguments{
\item{longterm_predictions}{}

\item{end_year}{}
}
\description{
Title
}
