% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_lm.R
\name{long_term_lm}
\alias{long_term_lm}
\title{Title}
\usage{
long_term_lm(longterm_all_data, test_set_steps = 2, testquant = 500)
}
\arguments{
\item{testquant}{}
}
\description{
Title
}
\examples{
working_directory <- getwd()
setwd(tempdir())
#longterm_model_data_example <- long_term_lm(longterm_all_data_example,test_set_steps=2,testquant = 500)
setwd(working_directory)
}
