% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_lm.R
\name{long_term_lm}
\alias{long_term_lm}
\title{Long-term forecast}
\usage{
long_term_lm(longterm_and_macro_data, test_set_steps = 2, testquant = 500)
}
\arguments{
\item{longterm_and_macro_data}{Dataframe. Containing the load data and macroeconomic indicators derived from \code{\link{get_macro_economic_data}}.}

\item{test_set_steps}{Integer. Number of time periods in the test set.}

\item{testquant}{Integer. Determines how many of the best ranked models are evaluated with cross validation.}
}
\value{
A dataframe with the input data and additional columns for test_set_steps and the for best three models longterm_model_predictions1, longterm_model_predictions2 and longterm_model_predictions3.
The dataset and the models are saved in the respective folder for the country.
}
\description{
This function predicts the long-term load data based on the provided time series and a set of macroeconomic variables. The model corresponds to
}
\details{
\deqn{\bar{D}_L(t_L)=\beta_{L,1}+\beta_{L,2}x_1(t_L)+...+ \beta_{L,10}x_{10}(t_L) \epsilon_L(t_L).}

where the covariates correspond to the loaded macroeconomic variables from \code{\link{get_macro_economic_data}}. For all possible covariate combinations, the three best models are chosen and stored. The included regressors are stored and the predicted and actual time series plotted.
}
\examples{
working_directory <- getwd()
setwd(tempdir())
example_longterm_predictions <- long_term_lm(example_longterm_and_macro_data,
test_set_steps=2,testquant = 500)
suppressMessages(
 unlink("./FR", recursive = TRUE, force = TRUE)
 )
setwd(working_directory)
}
\seealso{
See also function \code{\link{mid_term_lm}} and \code{\link{short_term_lm}} for the other prediction models and \code{\link{get_macro_economic_data}} for the covariate download.
}
