% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_term_future.R
\name{mid_term_future}
\alias{mid_term_future}
\title{Generate future mid-term demand predictions}
\usage{
mid_term_future(midterm_predictions, end_year, Tref = 18)
}
\arguments{
\item{midterm_predictions}{Dataframe. Generated by \code{\link{mid_term_lm}}}

\item{end_year}{Integer. Specifies the final year for which future predictions will be generated.}

\item{Tref}{Numeric. Reference temperature as basis for the calculation of cooling and heating days.}
}
\value{
The extended initial dataframe with the future predictions for the mid term model.
}
\description{
This function extends the mid-term demand predictions generated by \code{\link{mid_term_lm}} until a specified future year.
The unknown temperature-based covariates for future days are obtained by averaging over the past 3 years of the dataset.
The function also produces and saves visualizations of the actual and the predicted demand over the training, test, and future periods.
}
\examples{
working_directory <- getwd()
setwd(tempdir())
example_midterm_future_predictions <- mid_term_future(example_midterm_predictions,
end_year=2028,Tref=18)
suppressMessages(
 unlink("./FR", recursive = TRUE, force = TRUE)
 )
setwd(working_directory)
}
\seealso{
See also function \code{\link{long_term_future}} and \code{\link{short_term_future}} for the other prediction models.
}
