% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_term_future.R
\name{mid_term_future}
\alias{mid_term_future}
\title{Title}
\usage{
mid_term_future(midterm_predictions, end_year, Tref = 18)
}
\arguments{
\item{midterm_predictions}{Dataframe. Generated by \code{\link{mid_term_lm}}}

\item{end_year}{Integer. Specifies the final year for which future predictions will be generated.}

\item{Tref}{Numeric. Reference temperature as basis for the calculation of cooling and heating days.}
}
\value{
The extended initial dataframe with the future predictions for the mid term model.
}
\description{
Title
}
\examples{
\dontrun{
midterm_future_predictions <- mid_term_future(midterm_future_data_example,end_year=2028,Tref=19)
}
}
\seealso{
See also function \code{\link{long_term_future}} and \code{\link{short_term_future}} for the other prediction models.
}
