% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_term_lm.R
\name{mid_term_lm}
\alias{mid_term_lm}
\title{Mid-term forecast}
\usage{
mid_term_lm(
  demand_and_weather_data,
  Tref = 18,
  test_set_steps = 730,
  method = "temperature transformation"
)
}
\arguments{
\item{demand_and_weather_data}{Dataframe. Containing the mid-term load data, the holidays and weather data obtained from \code{\link{get_weather_data}}.}

\item{Tref}{Numeric. Reference temperature as basis for the calculation of cooling and heating days.}

\item{test_set_steps}{Integer. Number of time periods in the test set.}

\item{method}{String. Indicates which model selection process is used. If method="temperature transformation", the temperature values are transformed to heating and cooling
degree days to capture the non-linear relationship of temperature and electricity demand. If the method is set to "spline" a spline regression is instead used without
the transformation of the temperature data.}
}
\value{
The forecast of the best model fit is stored and the results are displayed in a plot.
}
\description{
The mid-term load series is forecasted based on the provided load time series and weather data. The prediction is either based on the (lagged) temperature data in combination with transformed variables for heating and cooling days or on a spline regression applied on the temperature data to account for non-linear effects.
}
\examples{

working_directory <- getwd()
setwd(tempdir())
example_midterm_predictions <- mid_term_lm(example_midterm_demand_and_weather_data$demand,
Tref=18, test_set_steps=730, method="temperature transformation")
suppressMessages(
 unlink("./FR", recursive = TRUE, force = TRUE)
 )
setwd(working_directory)

}
