% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_future.R
\name{short_term_future}
\alias{short_term_future}
\title{Title}
\usage{
short_term_future(shortterm_predictions, end_year)
}
\arguments{
\item{end_year}{Integer. Specifies the final year for which future predictions will be generated}
}
\description{
Title
}
