% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_future.R
\name{short_term_future}
\alias{short_term_future}
\title{Generate future short-term demand predictions}
\usage{
short_term_future(shortterm_predictions, end_year)
}
\arguments{
\item{shortterm_predictions}{Dataframe. Generated by \code{\link{short_term_lm}}}

\item{end_year}{Integer. Specifies the final year for which future predictions will be generated}
}
\value{
The extended initial dataframe with the future predictions for the short term model.
}
\description{
This function extends the short-term demand predictions generated by \code{\link{short_term_lm}} until a specified future year.
The function also produces and saves visualizations of the actual and the predicted demand over the training, test, and future periods.
}
\examples{
example_shortterm_future_predictions <- short_term_future(example_shortterm_predictions,
  end_year = 2028
)
}
\seealso{
See also function \code{\link{long_term_future}} and \code{\link{mid_term_future}} for the other prediction models.
}
