% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_lm.R
\name{short_term_lm}
\alias{short_term_lm}
\title{Title}
\usage{
short_term_lm(short_term_data, training_set_ratio = 0.2)
}
\arguments{
\item{training_set_ratio}{}
}
\description{
Title
}
\examples{
working_directory <- getwd()
setwd(tempdir())
shortterm_model_data_example <- short_term_lm(shortterm_holidays_example)
setwd(working_directory)
}
