% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_lm.R
\name{short_term_lm}
\alias{short_term_lm}
\title{Short-term forecast}
\usage{
short_term_lm(shortterm_demand_data, test_set_steps = 17520)
}
\arguments{
\item{shortterm_demand_data}{Dataframe. Containing the short-term load data from \code{\link{decompose_load_data}} and the added holiday dummy resulting from \code{\link{add_holidays_short_term}}.}

\item{test_set_steps}{Integer. Number of hours used for the test set. The default value of 17520 equals two years (2 * 8760 hours).}
}
\value{
A dataframe containing the original data series and the predicted data series.
}
\description{
The short-term load series is forecasted based on the provided hourly load data.
}
\examples{
working_directory <- getwd()
setwd(tempdir())
example_shortterm_predictions <- short_term_lm(example_shortterm_demand_data)
suppressMessages(
 unlink("./FR", recursive = TRUE, force = TRUE)
 )
setwd(working_directory)
}
