% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_lm.R
\name{short_term_lm}
\alias{short_term_lm}
\title{Short-term forecast}
\usage{
short_term_lm(short_term_data, test_set_steps = 17520)
}
\arguments{
\item{short_term_data}{Dataframe. Containing the short-term load data and the added holiday dummy resulting from \code{\link{add_holidays_short_term}}.}

\item{test_set_steps}{Integer. Number of time periods in the test set.}
}
\value{
A dataframe containing the original data series and the predicted data series.
}
\description{
The short-term load series is forecasted based on the provided hourly load data.
}
\examples{
working_directory <- getwd()
setwd(tempdir())
shortterm_model_data_example <- short_term_lm(shortterm_holidays_example)
setwd(working_directory)
}
